requirements.txt
alembic==1.4.2
aliyun-python-sdk-core==2.13.24
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-ecs==4.19.9
aliyun-python-sdk-kms==2.11.0
amqp==2.6.1
aniso8601==8.0.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
crcmod==1.7
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
flask-paginate==0.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
Mako==1.1.2
MarkupSafe==1.1.1
oss2==2.12.1
paginate==0.5.6
Pillow==7.2.0
pycryptodome==3.9.8
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python-memcached==1.59
pytz==2020.1
qiniu==7.2.10
redis==3.5.3
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.1
